src/requirements.txt
nltk==3.7
pandas==1.4.4
PyYAML==6.0
spacy==3.4.1
tensorflow==2.9.2
tensorflow_datasets==4.6.0
tensorflow_hub==0.12.0
tensorflow_text==2.9.0
textblob==0.17.1
typing_extensions==4.3.0
